refactor: add catrItem component
