+++ src/pages/cart/Cart.tsx
@@ -2,16 +2,15 @@ import { useEffect, useState } from 'react';
 
 import './Cart.scss';
 import { Link } from 'react-router-dom';
-import { Minus, Plus, ShoppingCart, Tag, Trash2 } from 'lucide-react';
+import { ShoppingCart, Tag, Trash2 } from 'lucide-react';
 import { useGameStore } from '../../store/store';
 import createCart from '../../utils/cart/createCart';
 import getCart from '../../utils/cart/getCart';
 import { Cart, ErrorObject, LineItem } from '@commercetools/platform-sdk';
-import updateItemQuantity from '../../utils/cart/updateItemQuantity';
-import removeItemFromCart from '../../utils/cart/removeItemFromCart';
 
 import apiRoot from '../../utils/sdkClient';
 import { ctpClient } from '../../utils/BuildClient';
+import { CartItem } from '../../components/basket/CartItem';
 
 export default function CartPage() {
     const {
@@ -25,8 +24,7 @@ export default function CartPage() {
         showStandardErrorMessage,
     } = useGameStore();
     const [cartItems, setCartItems] = useState<LineItem[]>([]);
-    const [isDeleteButtonDisabled, setDeleteButtonDisabled] = useState(false);
-    const [isCountButtonsDisabled, setCountButtonsDisabled] = useState(false);
+
     const [isRemovePromoButtonDisabled, setRemovePromoButtonDisabled] =
         useState(false);
     const [isApplyPromoButtonDisabled, setApplyPromoButtonDisabled] =
@@ -112,43 +110,6 @@ export default function CartPage() {
         setDiscount(subtotal - totalPrise);
         setTotal(totalPrise);
         setSubtotal(subtotal);
-    }
-
-    function removeItem(id: string): void {
-        setDeleteButtonDisabled(true);
-        removeItemFromCart(id, cartId!, cartVersion!)
-            .then((data) => {
-                if (data) {
-                    setCardVersion(data.version);
-                    showSuccessMessage(
-                        'The game has been successfully removed from the cart.'
-                    );
-                }
-            })
-            .catch(() => {
-                showStandardErrorMessage();
-            })
-            .finally(() => {
-                setTimeout(() => {
-                    setDeleteButtonDisabled(false);
-                }, 1500);
-            });
-    }
-
-    function updateQuantity(id: string, quantity: number): void {
-        setCountButtonsDisabled(true);
-        updateItemQuantity(id, cartId!, cartVersion!, quantity)
-            .then((data) => {
-                if (data) {
-                    setCardVersion(data.version);
-                }
-            })
-            .catch(() => {
-                showStandardErrorMessage();
-            })
-            .finally(() => {
-                setCountButtonsDisabled(false);
-            });
     }
 
     function clearCart() {
@@ -262,157 +223,7 @@ export default function CartPage() {
                 <div className={'cartLayout'}>
                     <div className={'cartItemsSection'}>
                         {cartItems.map((game) => (
-                            <div key={game.id} className={'cartItem'}>
-                                <div className={'itemContent'}>
-                                    {game.variant.images && (
-                                        <img
-                                            src={game.variant.images[0].url}
-                                            alt={game.name['en-US']}
-                                            className={'itemImage'}
-                                        />
-                                    )}
-
-                                    <div className={'itemDetails'}>
-                                        <h3 className={'itemTitle'}>
-                                            {game.name['en-US']}
-                                        </h3>
-                                        <div className={'itemPricing'}>
-                                            {game.price.discounted?.value
-                                                .centAmount ? (
-                                                <>
-                                                    <span
-                                                        className={
-                                                            'currentPrice'
-                                                        }
-                                                    >
-                                                        $
-                                                        {(
-                                                            Number(
-                                                                game.price
-                                                                    .discounted
-                                                                    ?.value
-                                                                    .centAmount
-                                                            ) / 100
-                                                        ).toFixed(2)}
-                                                    </span>
-                                                    <span
-                                                        className={
-                                                            'originalPrice'
-                                                        }
-                                                    >
-                                                        $
-                                                        {Number(
-                                                            game.price.value
-                                                                .centAmount
-                                                        ) / 100}
-                                                    </span>
-                                                </>
-                                            ) : (
-                                                <span
-                                                    className={'currentPrice'}
-                                                >
-                                                    $
-                                                    {(
-                                                        Number(
-                                                            game.price.value
-                                                                .centAmount
-                                                        ) / 100
-                                                    ).toFixed(2)}
-                                                </span>
-                                            )}
-                                        </div>
-
-                                        <div className={'itemActions'}>
-                                            <div className={'quantityControls'}>
-                                                <button
-                                                    onClick={() =>
-                                                        updateQuantity(
-                                                            game.id,
-                                                            game.quantity - 1
-                                                        )
-                                                    }
-                                                    className={'quantityButton'}
-                                                    disabled={
-                                                        game.quantity <= 1 ||
-                                                        isCountButtonsDisabled
-                                                    }
-                                                >
-                                                    <Minus />
-                                                </button>
-
-                                                <span
-                                                    className={
-                                                        'quantityDisplay'
-                                                    }
-                                                >
-                                                    {game.quantity}
-                                                </span>
-
-                                                <button
-                                                    onClick={() =>
-                                                        updateQuantity(
-                                                            game.id,
-                                                            game.quantity + 1
-                                                        )
-                                                    }
-                                                    className={'quantityButton'}
-                                                    disabled={
-                                                        isCountButtonsDisabled
-                                                    }
-                                                >
-                                                    <Plus />
-                                                </button>
-                                                <div className="totalGamePrice">
-                                                    <span
-                                                        className={
-                                                            'currentPrice'
-                                                        }
-                                                    >
-                                                        Total: $
-                                                        {game.price.discounted
-                                                            ?.value.centAmount
-                                                            ? (
-                                                                  (Number(
-                                                                      game.price
-                                                                          .discounted
-                                                                          ?.value
-                                                                          .centAmount
-                                                                  ) *
-                                                                      Number(
-                                                                          game.quantity
-                                                                      )) /
-                                                                  100
-                                                              ).toFixed(2)
-                                                            : (
-                                                                  (Number(
-                                                                      game.price
-                                                                          .value
-                                                                          .centAmount
-                                                                  ) *
-                                                                      Number(
-                                                                          game.quantity
-                                                                      )) /
-                                                                  100
-                                                              ).toFixed(2)}
-                                                    </span>
-                                                </div>
-                                            </div>
-
-                                            <button
-                                                onClick={() =>
-                                                    removeItem(game.id)
-                                                }
-                                                className={'removeButton'}
-                                                disabled={
-                                                    isDeleteButtonDisabled
-                                                }
-                                            >
-                                                <Trash2 />
-                                            </button>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
+                            <CartItem key={game.id} {...game} />
                         ))}
                     </div>
